+++ sdk/storage/azure-storage-common/inc/azure/storage/common/crypt.hpp
@@ -5,11 +5,16 @@
 
 #include <cstdint>
 #include <string>
+#include <vector>
 
 namespace Azure { namespace Storage {
 
-  std::string Base64Encode(const std::string& text);
-  std::string Base64Decode(const std::string& text);
+  std::string Base64Encode(const std::vector<uint8_t>& data);
+  inline std::string Base64Encode(const std::string& text)
+  {
+    return Base64Encode(std::vector<uint8_t>(text.begin(), text.end()));
+  }
+  std::vector<uint8_t> Base64Decode(const std::string& text);
 
   class Md5 {
   public:
@@ -18,16 +23,16 @@ namespace Azure { namespace Storage {
 
     void Update(const uint8_t* data, std::size_t length);
 
-    std::string Digest() const;
+    std::vector<uint8_t> Digest() const;
 
-    static std::string Hash(const uint8_t* data, std::size_t length)
+    static std::vector<uint8_t> Hash(const uint8_t* data, std::size_t length)
     {
       Md5 instance;
       instance.Update(data, length);
       return instance.Digest();
     }
 
-    static std::string Hash(const std::string& data)
+    static std::vector<uint8_t> Hash(const std::string& data)
     {
       return Hash(reinterpret_cast<const uint8_t*>(data.data()), data.length());
     }
@@ -41,16 +46,16 @@ namespace Azure { namespace Storage {
     void Update(const uint8_t* data, std::size_t length);
     void Concatenate(const Crc64& other);
 
-    std::string Digest() const;
+    std::vector<uint8_t> Digest() const;
 
-    static std::string Hash(const uint8_t* data, std::size_t length)
+    static std::vector<uint8_t> Hash(const uint8_t* data, std::size_t length)
     {
       Crc64 instance;
       instance.Update(data, length);
       return instance.Digest();
     }
 
-    static std::string Hash(const std::string& data)
+    static std::vector<uint8_t> Hash(const std::string& data)
     {
       return Hash(reinterpret_cast<const uint8_t*>(data.data()), data.length());
     }
@@ -61,8 +66,10 @@ namespace Azure { namespace Storage {
   };
 
   namespace Details {
-    std::string Sha256(const std::string& text);
-    std::string HmacSha256(const std::string& text, const std::string& key);
+    std::vector<uint8_t> Sha256(const std::vector<uint8_t>& data);
+    std::vector<uint8_t> HmacSha256(
+        const std::vector<uint8_t>& data,
+        const std::vector<uint8_t>& key);
     std::string UrlEncodeQueryParameter(const std::string& value);
     std::string UrlEncodePath(const std::string& value);
   } // namespace Details