+++ sdk/storage/azure-storage-common/src/shared_key_policy.cpp
@@ -18,18 +18,17 @@ namespace Azure { namespace Storage { namespace Details {
     string_to_sign += Azure::Core::Http::HttpMethodToString(request.GetMethod()) + "\n";
 
     const auto& headers = request.GetHeaders();
-    for (std::string headerName :
-         {"Content-Encoding",
-          "Content-Language",
-          "Content-Length",
-          "Content-MD5",
-          "Content-Type",
-          "Date",
-          "If-Modified-Since",
-          "If-Match",
-          "If-None-Match",
-          "If-Unmodified-Since",
-          "Range"})
+    for (std::string headerName : {"Content-Encoding",
+                                   "Content-Language",
+                                   "Content-Length",
+                                   "Content-MD5",
+                                   "Content-Type",
+                                   "Date",
+                                   "If-Modified-Since",
+                                   "If-Match",
+                                   "If-None-Match",
+                                   "If-Unmodified-Since",
+                                   "Range"})
     {
       auto ite = headers.find(Azure::Core::Strings::ToLower(headerName));
       if (ite != headers.end())
@@ -80,7 +79,8 @@ namespace Azure { namespace Storage { namespace Details {
     // remove last linebreak
     string_to_sign.pop_back();
 
-    return Base64Encode(
-        Details::HmacSha256(string_to_sign, Base64Decode(m_credential->GetAccountKey())));
+    return Base64Encode(Details::HmacSha256(
+        std::vector<uint8_t>(string_to_sign.begin(), string_to_sign.end()),
+        Base64Decode(m_credential->GetAccountKey())));
   }
 }}} // namespace Azure::Storage::Details